fix: Fix update books method
Previous method included a bug where if a book currently on a
shelf was updated to a new shelf it would be added again and
appear on 2 shelves.

The method to check if the book is already in the state.books
array now checks for the book ID rather than the book object.